Support type representations in general (#398)
### What

This PR adds support for type representations for base types and domain
types, building on top of the existing support for type representation
hinting of enum types.

Note that the wide numeric types default to numeric representation, see
the comment in the code. This will be relayed in the user docs in a
follow-up PR.

### How

The introspection query is seeded with a list of type representations
for base types, which it will use to build the final list of
represenations for base types, domain types, and enum types.