Use references more often to avoid cloning. (#403)
### What

Often, we accept parameters by value when a reference would do. This
means the caller might have to unnecessarily clone the value.

Clippy helped me figure out where this was unnecessary.

This changeset was brought to you by waiting for a meeting to start.

### How

I enabled some pedantic Clippy warnings and followed the trail.