ndc-spec 0.1.2 (#408)
### What

Support ndc-spec v0.1.2.
This includes changing the type representations of some types.

### How

The type representation has changed. We try to keep this change
backwards compatible by mapping deprecated variants (Number, Integer) to
new variants, and update types to more specific reps that didn't exist
before.

We also have a problem with representing int8 now. ndc-spec v0.1.2
defines Int64 to be represented as a json string, which is not what we
do. So we either have to lie (say it's an Int32 when it isn't) or break
compatibility and cast int8 to string.
In this PR I did the former, and I'll introduce the latter change next.