Support nested fields of array type (#409)
### What

This PR adds support for selecting nested fields that arise from
composite arrays.

### How

This builds on the earlier work for nested field selection support for
objects.

The main change is that the lateral joins that project the nested field
have to be nested rather than adjacent as well, since there are multiple
rows in the case of arrays.

The generated SQL is now of the form:
```
LEFT OUTER JOIN LATERAL (
  SELECT
    <collect_expression> AS "collected"
  FROM
    (
      < select of translate_fields(fields, nested_field_binding_alias, ... )
        which includes joins from recursive calls>
       FROM
         (
            SELECT
             (<field_binding_expression>).*
         ) AS <nested_field_binding> ON ('true')
     ) AS <nested_fields>
) AS <nested_fields_collect> ON ('true')
```

---------

Co-authored-by: Daniel Harvey <danieljamesharvey@gmail.com>