bugfix: allow aliasing returning and affected_rows fields (#407)
### What

It should be possible to alias the returning and affected_rows fields
when running a mutation.
This is doubly important because this is what the cli/lsp/engine do by
default for `affected_rows` :)

### How

Instead of having the alias satisfy the field name, we ask the Field to
do that.
We traverse the result fields, find the ones named "returning" and
"affected_rows", and use that, instead of trying to fetch by the alias
name in the fields map.