Separate business and exchange types (#419)
### What

We want to be able to confidently evolve the configuration format
without risk of accidentally changing the behavior of previous versions.

This PR has no user-observable change in behavior.

Introducing a new major version of the configuration format is now a
much more well-defined task, which should consist most of just
duplicating the entire `version3` module hierarchy and then introducing
the new changes in the duplicated copy.

### How

In order to achieve this we duplicate the complete metadata type
hierarchy into configuration crate and introduce functions that convert
between the now separate exchange types and business types, and ensure
that all logic that has to do with serialization only lives with the
exchange types (the version types)