remove transactions in queries (#421)
### What

Running queries in transactions require that we run 3 statements: begin,
the query, commit.
We use transactions to enforce a `READ ONLY` setting so that native
queries which run mutations will be rejected.
However, each roundtrip takes time. If we can avoid it, it would improve
the performance of queries.

In the PR we use the fact the mutating sql statements can only occur in
a CTE at the top-level, and wrap native queries in another layer of CTE,
thereby enforce the read only requirement without using transactions.

### How

1. We remove the `begin` and `commit` statements before/after a query.
2. We wrap each query in another CTE of the form `WITH
<native_query_name> as (WITH <alias> as (<native_query>) SELECT * FROM
CTE as <alias>) SELECT ...`
3. We extract the global table alias index from the state and use it in
wrapping the CTEs.
4. We wrap ctes in the caller side because this code is used for
mutations as well, for which we don't want to apply this hack.