cast int64 and numeric to string by default (#416)
### What

We want to be able to stringify fields of types `int8` and `numeric` in
the query response so that we have better interoperability with
javascript clients.

This PR changes the default type representation of `int8` and `numeric`
types (maintaining the existing type representation if one exists) and
stringify numbers.

This option can be change via the configuration by editing the mapping
in the `metadata.typeRepresentations`:

```diff
  { ...
    "metadata": {
      ...
      "typeRepresentations": {
        "Phone": "string",
        "CardSuit": {
          "enum": ["hearts", "clubs", "diamonds", "spades"]
        },
        ...
-       "int8": "int64",
+       "int8": "int64AsString",
-       "numeric": "bigDecimal",
+       "numeric": "bigDecimalAsString",
        ...
      }
      ...
    }
  }
```

### How

1. We introduce new type representations: `Int64AsString`,
`BigDecimalAsString`.
2. Make sure we use the existing type representation configuration and
override the default if it exists.
3. In field selection after getting an expression, we cast it if the
type representation requires it. (For arrays, we cast to `cast_type[]`)
4. For composite types we unpack them if the don't specify fields, and
cast required fields recursively.