Add schema names to composite types.
Also translate composite types into sql::ast::ScalarTypeName using their
name and schema rather than the ndc-schema object type name, which had
incidentally been the same as the (unqualified) type name.

This means need access to the Env when translating values/variables of
composite type. Most of the files changed in this PR is a result from
knock-on effects of this.