Add schema names to composite types. (#432)
### What

This PR adds schema names to composite types in metadata. Schema names
are not introspected yet, but this will happen in a subsequent PR.

Also translate composite types into sql::ast::ScalarTypeName using their
name and schema rather than the ndc-schema object type name, which had
incidentally been the same as the (unqualified) type name.

This means we need access to the `Env` when translating values/variables
of
composite type. Most of the files changed in this PR is a result from
knock-on effects of this.

This PR is a follow-up to
https://github.com/hasura/ndc-postgres/pull/431.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>