Use the full SHA when updating ndc-postgres-multitenant.
Cargo chokes when using different variations of Git commit refs, even if
they eventually point to the same commit. It believes they're different
versions and won't reconcile them, as there's no semver-based versioning
to lean on.

It's safer to always use the full SHA in order to avoid the situation
where different SHA lengths are used.