Avoid `format!` and `.as_str()` where possible.
Sometimes there are simpler mechanisms for constructing and referencing
strings.

Where the result of `format!` has been passed to `push_str`, we can call
that multiple times instead, which IMO makes things clearer and avoids
building an intermediate string.

When we use it to stringify a single value, we can call `.to_string()`
instead.

Paths can be built using `Path::join`.

`.as_ref()` can be replaced with `&` in many places.