ScalarTypeName ast: use enums instead of flags (#448)
### What

Followup to the review suggestion:
https://github.com/hasura/ndc-postgres/pull/431#discussion_r1571231393

Makes the AST for ScalarType a bit more structured.
Using enums instead of flags makes things more explicit, so if we make a
mistake it would not be one of "I forgot to set the flag" but rather "I
explicitly chose the wrong variant".

### How

Use the following representation:

```rust
pub enum ScalarType {
    BaseType(ScalarTypeName),
    ArrayType(ScalarTypeName),
}

pub enum ScalarTypeName {
    Qualified {
        schema_name: SchemaName,
        type_name: String,
    },
    Unqualified(String),
}
```

Instead of the previous:

```rust
pub struct ScalarTypeName {
    pub type_name: String,
    pub schema_name: Option<SchemaName>,
    pub is_array: bool,
}
```