Introduce configuration version 4: Support schema-qualified type (#444)
### What

This PR introduces the ability to refer to scalar types and complex
types by a separate name in the ndc schema. This name is formed taking
into account the schema the type is defined in, in the same manner as we
treat tables.

The implementation is limited so far as it only includes the parsing of
this format and not its construction (though doing so for testing is as
simple as changing the default empty RawConfiguration value).

The reason for this is that this is already a sizable PR, so it would be
good to review and integrate before it includes CLI support and
upgrading of tests.
 
### How

While this sounds innocuous it actually introduces an extra level of
indirection. We used to only refer to scalar types by their unqualified
name in both SQL generation and NDC schema. Now we record the defined
name and schema of a type and derive a new name based off these which we
use to refer to the type everywhere in the generated schema.

A knock-on effect of this is that it is now more appropriate to produce
a dedicated "ScalarTypes" object in introspection that includes both
name data and aggregation functions and comparison operators for scalar
types in nested objects. This actually simplifies the business logic of
the dataconnecter itself somewhat as well, since this new structure has
the same shape as the schema-response.