When acquiring a connection, ping the db only if it has been idle (#450)
### What

When we acquire a database connection from the pool, there's a high
chance that the connection is alive and well. We add a new pool setting
that will only ping the database after a certain amount of time the
connection has been idle (default 60 seconds). This settings can be
turned off.

### How

Adding a new pool settings option, and essentially doing this:
https://docs.rs/sqlx/latest/sqlx/postgres/type.PgPoolOptions.html#example-custom-test_before_acquire-logic