Add support for upgrading version 3 configurations to version 4 (#452)
### What

This PR adds the ability to upgrade a configuration directory in version
3 to an equivalent one in version 4.

The cli now supports the command `cli upgrade --from-dir <existing
configuration dir> --to-dir <new configuration dir>`, with the intended
semantics that it upgrades the configuration in `<existing configuration
dir>` to the newest supported version.

### How

Many places in the code (though mostly tests) is tightly coupled with
internal details of version 3 and will attempt to construct and
manipulate configurations directly.

This PR attempts to define a version-agnostic interface that can be used
instead.