Migrate tests to access configuration through public interface (#457)
### What

This PR ensures that all test code that relies on NDC configurations
does so via the public interface rather than by directly depending on a
specific configuration version format module.

The tests that have the version format as their subject matter have been
moved into the configuration crate, cohabiting with the code they test.

The tests that don't have the version format as their subject matter now
rely only on the public interface of the configuration crate. The
biggest consequence of this by diff volume is that what used to be
`tables.json` holding a `Metadata` value now becomes full
`configuration.json` files, with their original content in the
`metadata` section and a `version: "3"` tag as well.