Test that the version produced by `ndc-postgres-cli initialize` is stable (#472)
### What

This PR adds a test that asserts that the configuration version produced
by the cli `initialize` command hasn't changed.

### How

The test runs the initialize command and asserts on the value of the
`version` key in `configuration.json`. This is a somewhat simplified
model of version checking, but it holds for the ones we have currently.