rename ast:: to sql::ast:: (#476)
### What

The module hierarchy is built so that the `sql::ast::` module is
imported qualified in this fashion so there's no ambiguity when one runs
into a type representing part of the SQL AST.

For example, `ast::Expression` does not describe whether this ast is
from ndc-spec, an internal ndc-postgres IR, or SQL, while
`sql::ast::Expression` makes it perfectly clear.

This PR replaces the usage of `ast::` with `sql::ast::`.

### How

This is just a rename.