Insert multiple objects instead of a single object (#477)
### What

The currently generated insert implementation only provides a mechanism
to insert a single object / row via the `_object` parameter.
In this PR we replace the `_object` parameter with an `_objects`
parameter that accepts a list of objects, so we can insert multiple
objects / rows (bulk insert).

### How

1. Change the schema to an array of objects
2. Make the sql::ast::Insert have a vector of vector of values, instead
of just a vector of values, that might be Default as well as Expression
3. match user input as array of objects
4. Get a union of the columns from all objects, and traverse each object
and insert a Default when a column is missing