Support inserting an empty object (#478)
### What

We should support the scenario of inserting an empty object `{}`, which
means use the default for all columns.

### How

To support a table that has only generated columns such as this one:

```sql
create table t(x int generated always as (1) stored);
```

We need to generate the following SQL to insert N empty objects:

```sql
insert into t select from generate_series(1,N);
```

So in case all of the objects are empty, we translate to the query plan
above.