ndc-spec 0.1.3 with filter/order by nested fields (#486)
### What

1. Update ndc-spec to version
[0.1.3](https://github.com/hasura/ndc-spec/releases/tag/v0.1.3).
2. Implemented the new capabilities: filter and order by nested fields.

### How

#### Nested fields selection

We add SQL syntax for accessing an expression's field. The syntax is
`(<expression>).<field>`

#### Order by nested fields

We add the field selection to the sort expression in the `ORDER BY`
clause so it looks like this: `ORDER BY (column).field ASC`.
Implementation wise, we thread the field path like we do for the sort
direction, and wrap the expressions in a NestedFieldSelect clause.

#### Filter by nested fields

- We directly wrap columns that contain field paths when they are used
inside the predicate.
- We extract the type of a nested field by recursively looking it up in
the environment.