introduce unique_constraints module (#490)
### What

For Delete by key we have a function that returns the unique constraints
of a table.
This can be useful for Update by key, which we will introduce soon, as
well.

In this PR we move this function to its own module.

We also change a word in the schema description of delete.