Unify query part handling for rows and aggregates (#492)
### What

fixes https://hasurahq.atlassian.net/browse/PG-112

Previously, order by and limit clauses were not applied for the
aggregates part of the query as the spec said it was not needed,
however, that is a mistake which we now fix.

### How

To fix this, we unify the way we handle rows and aggregates.
Callers will no longer need to call `translate_rows_query` and
`translate_aggregates_query` separately and then try to stitch them
together, but rather call one part, `translate_query`, which will call
both.

Each translate part will call the same function, `translate_query_part`
which translates the common elements (where, order by, limit, etc.), but
will different in the way it handles fields or aggregates.

This way, we make sure that an ndc-spec query is translated consistently
between the two parts - rows and aggregates.