generate mutations for multi-column unique constraints (#500)
### What

A unique constraint on a table could be defined using multiple column.
For example in `PlaylistTrack` the pair `(PlaylistId, TrackId)` is a
unique constraint, and we will generate a procedure with the following
structure:

```js
experimental_delete_PlaylistTrack_by_PlaylistId_and_TrackId(
    PlaylistId: int,
    TrackId: int,
    filter: boolexpr
)
```

and a similar one for updates.

### How

- We change the mutation type to include multiple keys instead of a
single one.
- Instead of using `get_non_compound_uniqueness_constraints()`, we just
traverse the constraints list a generate add an argument per unique
constraint key.
- We generate cute names and descriptions by chaining the key names.
- We change the schema by iterating over the keys.
- We do the same for update and delete.
- We emit a warning and skip generating a procedure if something
unexpected happens.