Add configuration version metrics (#504)
### What

This PR adds a metric representing the configuration version that a
running ndc-postgres instance uses.

This will help us track adoption of new versions and inform deprecation
timelines and documentation.

### How

Rather than extending the existing metrics in the query_engine crate we
make a new Metrics struct in the ndc_postgres_configuration crate. This
is to prevent the query_engine metrics from accumulating unrelated cruft
and reduce the complexity of the module graph.