Fix schema generation for insert and update input objects (#514)
### What

For the input objects of mutations (`objects`, `update_columns`),
columns that have defaults should be marked as nullable, and columns
that are generated should not be listed (this one worked for insert but
not update).

### How

1. Moved the schema gen helpers that are only relevant for mutations
into the mutation directory
2. Moved the part that handles adding or not adding generated columns
out side of insert handling, and use it in both insert and update
3. Pattern match on the column info and specify the type according to
the rules in `What`.