Implement native query builder in the CLI plugin (#511)
### What

To improve the DX of native queries, we want a CLI interface with which
we can point the CLI at a SQL(ish - a native query) file and generate
the stanzas for V3 metadata. This PR sort of does it.

### Design

Introduce the following subcommands:

```
Commands on Native Operations

Usage: ndc-postgres-cli native-operation <COMMAND>

Commands:
  list    List the existing Native Operations
  create  Create a new Native Operation from a SQL file
  delete  Delete an existing Native Operation from the configuration
  help    Print this message or the help of the given subcommand(s)

Options:
  -h, --help  Print help
```

#### Create

```
Create a new Native Operation from a SQL file

Usage: ndc-postgres-cli native-operation create [OPTIONS] --operation-path <OPERATION_PATH> --kind <KIND>

Options:
      --operation-path <OPERATION_PATH>
          Relative path to the SQL file inside the connector configuration directory
      --kind <KIND>
          Operation kind [possible values: query, mutation]
      --override
          Override the Native Operation definition if it exists
  -h, --help
          Print help
```

#### Delete

```
Delete an existing Native Operation from the configuration

Usage: ndc-postgres-cli native-operation delete --name <NAME> --kind <KIND>

Options:
      --name <NAME>  The name of the Native Operation
      --kind <KIND>  Operation kind [possible values: query, mutation]
  -h, --help         Print help
```

#### List

```
List the existing Native Operations

Usage: ndc-postgres-cli native-operation list

Options:
  -h, --help  Print help
```

### How

The user supplies a file containing SQL(ish - a native operation) code,
and specifies the operation kind, and we use the sqlx `describe` method
to fetch information about the operation. Specifically, whether it
compiles or not, and what is the error if not, and what are the oids of
the types of the arguments and returning columns.

From this information, we construct a NativeQueryInfo and add it to the
configuration, then update it to fetch information about types we might
not have tracked in the configuration yet.

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>
Co-authored-by: Samir Talwar <samir.talwar@hasura.io>
Co-authored-by: Philip Lykke Carlsen <plcplc@gmail.com>