config v5 with native operations rename (#522)
### What

We introduce ndc-postgres configuration version "v5".

This version changes two things:

1. Renames the field `nativeQueries` to `nativeOperations`, and instead
of containing native queries and native mutations together with a flag
on each marking if it is mutation or not (`isProcedure`), we nest the in
different objects, so we have:

   ```
   nativeOperations:
     queries:
     mutations:
   ```

2. We nest `scalarTypes` and `compositeTypes` under one field, `types`:

   ```
   types:
     scalar:
     composite:
   ```


### How

Might be useful to read as commits.

We first copy paste v4 to v5, then create types for `Types` and
`NativeOperations` which nest the other information, then we fix all the
relevant things: ndc-spec schema, cli handling, upgrading from older
configs, conversion to runtime config, etc.

Then we fix the tests, including introducing new configs and changing
the paths the the configs.

Then we get rid of the older configurations in the `static/<db>`
directories, as they are not longer used for anything.