Array elements are always nullable (#571)
### What

PR #565 fixed field nullability of composite type fields, but neglected
to treat the similar issue that exists for arrays.

In Postgres, any array type (e.g. `int4[]`, `text[]`, etc.) may contain
elements which are null. The type system does not support declaring
arrays with non-null elements.

This PR fixes that issue.

### How

Whenever we have to translate an ndc-postgres array type to an ndc-spec
array type in the schema generation we annotate the element type as
being nullable.