implemented filter by nested collection (#573)
### What

Implementing an upcoming feature in ndc-spec:
https://github.com/hasura/ndc-spec/pull/166

We want to support filtering by a field inside an array column.

For example: I want to get all institutions that where there exists a
person who's last name is "Hughes" among their staff.

This requires an update to ndc-spec and ndc-sdk-rs.

### How

When we get a request with an exists over a nested field collection, we
construct a nested field selection expression, unnest the results, and
use that as our new "current table".