Automatically create PRs to bump versions in ndc-postgres (#594)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

<!-- Consider: do we need to add a changelog entry? -->

This PR adds a `bump-version.yaml` workflow that can be manually
triggered in the `Actions` tab. When triggered and given a new version
of `ndc-postgres`, a PR will be generated that bumps `ndc-postgres` to
that version. In practice, this means:

* Updating the `Cargo.toml` file and associated lock file.
* Moving everything in `changelog.md` from `Unreleased` to a section
with the appropriate version title, and creating a new, blank
`Unreleased` section.

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->
GitHub actions, including a fancy new action we found to create pull
requests against the repo.