Fix invalid joins when the order by relationship predicate is null (#655)
### What
If a relationship used in an order by specified a null predicate, the
SQL generator would fail to add the join conditions when joining the
related table into the query. This would result in an incorrect cross
product between the table and the related table.

### How

If the relationship predicate is null, we now correctly add the table's
join conditions.

Existing tests that assumed all predicates would contain an always true
expression have been changed to contain a null predicate and an extra
test has been added to cover receiving an always true predicate (to make
sure that still works).

Fixes ENG-1386